Add CSS to the Index HTML file
to complete TOP CSS Exercise 02
My original css code is visible in Style2.css
while my revised code is present in the main Style.css file.
The revised code in Style.css was created by consulting
the Odin Project solution files and working through it's
decisions line by line to compare my code against the
best practices and to solve for my failure to shrink the
gap at the top from 16px to 8px. Learned a lot on this one!